drivers: ethernet: stm32: H7 Ethernet support test
This driver is being tested to use Ethernet on H7.
F7 and H7 have a slightly different register configuration.

The first goal was the occurrence of the receive interrupt function HAL_ETH_RxCpltCallback().
However, it is not currently working properly.

The register of the PHY chip is read normally.

Because it is a test source, the source is dirty.

There may be a bug when using H7 Ethernet.
1. Bug in ETH_DMARxDescListInit - reported in the forum
:https://github.com/STMicroelectronics/STM32CubeH7/issues/7
2. [bug fixes] STM32H7 Ethernet
:https://community.st.com/s/question/0D50X0000C6eNNSSQ2/bug-fixes-stm32h7-ethernet?t=1581554010199&searchQuery=
3. FAQ: Ethernet not working on STM32H7x3
:https://community.st.com/s/article/FAQ-Ethernet-not-working-on-STM32H7x3